tests: move to jest
- also fixed security deps issue
