+++ lib/geo-from-ip.js
@@ -1,9 +1,8 @@
-'use strict'
-
 // get node packages
 const debug = require('debug')('geo-from-ip:main')
 const path = require('path')
 const mmdbreader = require('maxmind-db-reader')
+
 const city = mmdbreader.openSync(
   path.resolve(path.join(__dirname, '/../mmdb/GeoLite2-City.mmdb'))
 )
@@ -18,8 +17,8 @@ debug(`geo-from-ip ready for action!`)
 function allData (ip) {
   debug(`allData function called with ip: ${ip}`)
   // get all data for ip
-  let geodata = city.getGeoDataSync(ip)
-  let result = {
+  const geodata = city.getGeoDataSync(ip)
+  const result = {
     code: {}
   }
