+++ lib/update.js
@@ -21,7 +21,7 @@ function isRemoteNewer (dest, remote, cb) {
     cb(null, false)
   } else {
     // if dest file is not a file, remove it
-    let stats = fs.statSync(dest)
+    const stats = fs.statSync(dest)
     if (!stats.isFile()) {
       debug(`${dest} is not a file`)
       fs.unlink(dest, () => {
@@ -72,7 +72,7 @@ function download (dest, remote, cb) {
  * @param  {Function} cb     true/false
  */
 function dl (item, cb) {
-  let dest = config.dbDir + '/' + item.local
+  const dest = config.dbDir + '/' + item.local
 
   isRemoteNewer(dest, item.remote, (err, newer) => {
     if (err) {
