+++ package.json
@@ -1,11 +1,12 @@
 {
   "name": "geo-from-ip",
-  "version": "1.2.1",
+  "version": "1.2.2",
   "description": "Get geolocation 🌐 information about an IP address 📲",
   "main": "lib/geo-from-ip.js",
   "scripts": {
-    "test": "standard --fix && DEBUG=geo-from-ip:* mocha",
-    "postinstall": "cross-env DEBUG=geo-from-ip:* node lib/update.js"
+    "fix": "standard --fix",
+    "postinstall": "cross-env DEBUG=geo-from-ip:* node lib/update.js",
+    "test": "DEBUG=geo-from-ip:* jest"
   },
   "repository": {
     "type": "git",
@@ -32,16 +33,13 @@
     "url": "https://github.com/VikramTiwari/geo-from-ip/issues"
   },
   "engines": {
-    "node": ">=6.10.0"
+    "node": ">=10"
   },
   "standard": {
     "globals": [
-      "after",
-      "afterEach",
-      "before",
-      "beforeEach",
       "describe",
-      "it"
+      "expect",
+      "test"
     ]
   },
   "homepage": "https://github.com/VikramTiwari/geo-from-ip",
@@ -53,10 +51,8 @@
     "request": "^2.88.0"
   },
   "devDependencies": {
-    "chai": "^4.2.0",
     "chance": "^1.0.18",
-    "expect": "^24.7.1",
-    "mocha": "^6.1.3",
-    "standard": "^12.0.1"
+    "jest": "^24.8.0",
+    "standard": "^13.1.0"
   }
 }
